move event_card_info.dart
